requirements_updated.txt
# Core dependencies
numpy==1.19.5
scipy==1.5.4
Pillow==8.4.0
scikit-image==0.17.2
scikit-learn==0.24.2
pandas==1.1.5
matplotlib==3.3.4
tqdm==4.64.1

# Deep Learning frameworks
torch==1.10.2
torchvision==0.11.3
tensorflow==1.15.0

# Adversarial attacks
advertorch==0.2.4

# Image processing
opencv-python==4.5.5.64
moviepy==1.0.3

# Data handling
lmdb==1.3.0
h5py==2.10.0
PyYAML==5.4.1

# Additional utilities
six==1.16.0
cryptography==3.4.8
Cython==0.29.32

# Optional (for Chainer-based models)
chainer==7.8.1
